+++ lib/appconfig.php
@@ -239,6 +239,20 @@ class AppConfig {
     public $_customization_logo = "customization_logo";
 
     /**
+     * The config key for the zoom
+     *
+     * @var string
+     */
+    public $_customization_zoom = "customization_zoom";
+
+    /**
+     * The config key for the autosave
+     *
+     * @var string
+     */
+    public $_customization_autosave = "customization_autosave";
+
+    /**
      * @param string $AppName - application name
      */
     public function __construct($AppName) {