+++ templates/editor.php
@@ -30,11 +30,19 @@
     style("onlyoffice", "editor");
     script("onlyoffice", "desktop");
     script("onlyoffice", "editor");
+    if (!empty($_["directToken"])) {
+        script("onlyoffice", "directeditor");
+    }
 ?>
 
 <div id="app">
 
-    <div id="iframeEditor" data-id="<?php p($_["fileId"]) ?>" data-path="<?php p($_["filePath"]) ?>" data-sharetoken="<?php p($_["shareToken"]) ?>"></div>
+    <div id="iframeEditor"
+        data-id="<?php p($_["fileId"]) ?>"
+        data-path="<?php p($_["filePath"]) ?>"
+        data-sharetoken="<?php p($_["shareToken"]) ?>"
+        data-directtoken="<?php p($_["directToken"]) ?>"
+        data-inframe="<?php p($_["inframe"]) ?>"></div>
 
     <?php if (!empty($_["documentServerUrl"])) { ?>
         <script nonce="<?php p(base64_encode($_["requesttoken"])) ?>"